Adjusted the display of invoices and Appointments
On the patient page the columns patient id, and modified were removed.
On the doctor page, the columns patient id, created, and modified were
removed.
Also it was changed so the user would see the patients name on the
doctors appointment section. Versus on the patient's page, the Doctors
name shows up.